Move `doRequest` in its own file
