+++ packages/plugins/rum/src/sdk.ts
@@ -2,7 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { doRequest } from '@dd/core/helpers';
+import { doRequest } from '@dd/core/helpers/request';
 import type { GlobalContext, InjectedValue } from '@dd/core/types';
 
 import type { RumOptionsWithDefaults, RumOptionsWithSdk } from './types';
